pyproject.toml
[tool.poetry]
name = "tbot_wind"
version = "0.1.0"
description = "A flexible trading bot for interacting with Interactive Brokers and sending signals via webhooks."
authors = ["Sangwook Lee <info@plusgenie.com>"]
license = "GPL-3.0-only OR Commercial"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
ib-async = "*"
pandas = "2.2.2"
numpy = "^1.21.0"
aiohttp = "^3.7.4"
loguru = "^0.5.3"
python-dotenv = "^0.19.0"
tabulate = "^0.9.0"
cachetools = "^5.4.0"
redis = "^5.0.8"
msgpack = "^1.0.8"
setuptools = "^73.0.1"

# Use the development branch of pandas-ta from GitHub
pandas-ta = {git = "https://github.com/twopirllc/pandas-ta.git", branch = "development"}
talipp = "^2.4.0"
yfinance = "^0.2.43"
openai = "^1.51.1"
pytest-asyncio = "^0.24.0"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
from = "src"
include = "tbot_wind"

[tool.poetry.scripts]
# Define a CLI command called "tbot-wind" that executes the main function
tbot-wind = "tbot_wind.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.black]
line-length = 79
